+++ src/data/events-data.js
@@ -25,28 +25,6 @@ const eventsData = {
   photosDefaultURL: `http://picsum.photos/300/150?r=`
 };
 
-// Подобрать приблизительно правдоподобную точку назначения согласно типу события.
-const getRandomDestination = (type, data) => {
-  switch (type) {
-    case `Flight`:
-      return getRandomElement(data.destination.cities);
-
-    case `Check-in`:
-      return getRandomElement(data.destination.checkinPoints);
-
-    case `Sightseeing`:
-      return getRandomElement(data.destination.sights);
-
-    case `Restaurant`:
-      return getRandomElement(data.destination.eatingPoints);
-
-    default:
-      return getRandomElement(new Set([...data.destination.cities,
-        ...data.destination.sights,
-        ...data.destination.eatingPoints,
-        ...data.destination.checkinPoints]));
-  }
-};
 
 const getOffer = (offerDescription, config) => ({
   description: offerDescription,
